Refactors the Filterinput Component
